museekd: load uint64 file size properly.
When loading the download list from disk at startup time, museekd
was incorrectly reading 64-bit integers.  This caused large files
to be marked (and displayed) as smaller than their true size and
prematurely stop downloading, leaving a partial file that was
marked as complete.  This patch fixes it, as well as two related
printf-style log message format directives.

+++ museekd/util.h
@@ -162,7 +162,7 @@ static inline int read_off(std::ifstream * ifs, uint64* r) {
 		return -1;
 
 	for(uint64 j = 0; j < 8; j++) {
-		(*r) += d[j] << (j * 8);
+		(*r) += (uint64)d[j] << (j * 8);
 	}
 	return 8;
 }
